--- bitwarden_license/src/Commercial.Core/SecretManager/Commands/AccessPolicies/CreateAccessPoliciesCommand.cs
@@ -1,45 +1,0 @@
-﻿using Bit.Core.Exceptions;
-using Bit.Core.SecretsManager.Commands.AccessPolicies.Interfaces;
-using Bit.Core.SecretsManager.Entities;
-using Bit.Core.SecretsManager.Repositories;
-
-namespace Bit.Commercial.Core.SecretManager.Commands.AccessPolicies;
-
-public class CreateAccessPoliciesCommand : ICreateAccessPoliciesCommand
-{
-    private readonly IAccessPolicyRepository _accessPolicyRepository;
-
-    public CreateAccessPoliciesCommand(IAccessPolicyRepository accessPolicyRepository)
-    {
-        _accessPolicyRepository = accessPolicyRepository;
-    }
-
-    public async Task<List<BaseAccessPolicy>> CreateAsync(List<BaseAccessPolicy> accessPolicies)
-    {
-        var distinctAccessPolicies = accessPolicies.DistinctBy(baseAccessPolicy =>
-        {
-            return baseAccessPolicy switch
-            {
-                UserProjectAccessPolicy ap => new Tuple<Guid?, Guid?>(ap.OrganizationUserId, ap.GrantedProjectId),
-                GroupProjectAccessPolicy ap => new Tuple<Guid?, Guid?>(ap.GroupId, ap.GrantedProjectId),
-                ServiceAccountProjectAccessPolicy ap => new Tuple<Guid?, Guid?>(ap.ServiceAccountId, ap.GrantedProjectId),
-                _ => throw new ArgumentException("Unsupported access policy type provided.", nameof(baseAccessPolicy))
-            };
-        }).ToList();
-
-        if (accessPolicies.Count != distinctAccessPolicies.Count)
-        {
-            throw new BadRequestException("Resources must be unique");
-        }
-
-        foreach (var accessPolicy in accessPolicies)
-        {
-            if (await _accessPolicyRepository.AccessPolicyExists(accessPolicy))
-            {
-                throw new BadRequestException("Resource already exists");
-            }
-        }
-
-        return await _accessPolicyRepository.CreateManyAsync(accessPolicies);
-    }
-}
