--- bitwarden_license/src/Commercial.Core/SecretManager/SecretsManagerServiceCollectionExtensions.cs
@@ -1,11 +1,0 @@
-﻿using Microsoft.Extensions.DependencyInjection;
-
-namespace Bit.Commercial.Core.SecretManager;
-
-public static class SecretsManagerServiceCollectionExtensions
-{
-    public static void AddCommercialSecretsManagerServices(this IServiceCollection services)
-    {
-        services.AddSecretManagerServices();
-    }
-}
