+++ bitwarden_license/src/Commercial.Core/SecretsManager/SecretsManagerServiceCollectionExtensions.cs
@@ -1,0 +1,11 @@
+﻿using Microsoft.Extensions.DependencyInjection;
+
+namespace Bit.Commercial.Core.SecretsManager;
+
+public static class SecretsManagerServiceCollectionExtensions
+{
+    public static void AddCommercialSecretsManagerServices(this IServiceCollection services)
+    {
+        services.AddSecretsManagerServices();
+    }
+}
