+++ bitwarden_license/test/Commercial.Core.Test/SecretsManager/ServiceAccounts/CreateServiceAccountCommandTests.cs
@@ -1,0 +1,25 @@
+﻿using Bit.Commercial.Core.SecretsManager.Commands.ServiceAccounts;
+using Bit.Core.SecretsManager.Entities;
+using Bit.Core.SecretsManager.Repositories;
+using Bit.Test.Common.AutoFixture;
+using Bit.Test.Common.AutoFixture.Attributes;
+using Bit.Test.Common.Helpers;
+using NSubstitute;
+using Xunit;
+
+namespace Bit.Commercial.Core.Test.SecretsManager.ServiceAccounts;
+
+[SutProviderCustomize]
+public class CreateServiceAccountCommandTests
+{
+    [Theory]
+    [BitAutoData]
+    public async Task CreateAsync_CallsCreate(ServiceAccount data,
+      SutProvider<CreateServiceAccountCommand> sutProvider)
+    {
+        await sutProvider.Sut.CreateAsync(data);
+
+        await sutProvider.GetDependency<IServiceAccountRepository>().Received(1)
+            .CreateAsync(Arg.Is(AssertHelper.AssertPropertyEqual(data)));
+    }
+}
